requirements.txt
python-telegram-bot>=20.0
pymongo>=4.0.0
python-dotenv>=1.0.0 